Updating versions
- Updated Quarkus from 2.11.1.Final to 2.14.1.Final.
- Changed dependencies from quarkus-resteasy-reactive to quarkus-resteasy (classic) as this example is not using reactive programming, so, it's better use the classic extension to be more clear.
- Updated version from Keycloak 18 Legacy (WildFly) to Keycloak 20.0.1 (Quarkus).
- Updated PostgreSQL from 13.3 to 15.1.
- Added Postman collection for test purposes.
