+++ README.md
@@ -3,7 +3,7 @@
 This project is a Java REST API application that was configured to use [Keycloak](https://www.keycloak.org) as access management.
 
 It was bootstrapped using [Quarkus](https://quarkus.io/) with the following extensions/dependencies:
-- RESTEasy reactive.
+- RESTEasy classic.
 - Hibernate ORM Panache and Hibernate Validator.
 - Liquibase (With PostgreSQL as the database).
 - OIDC to enable the integration with Keycloak.
@@ -25,6 +25,7 @@ This project follows the default Java structure `/src/main`.
 - `/src/main/docker` - The Docker files that were generated by Quarkus can be used to deploy the application depending on how you build the application (JVM, jar, or Native with GraalVM).
 - `/src/main/docker/keycloak`: The Docker files to build the Keycloak Docker image required for this project.
 - `/src/main/jenkins` - Jenkins file and resources that are used to set up the pipeline to build and deploy the application.
+- `/src/main/jenkins` - Postman collection to be imported for test purposes.
 
 ## Running tests
 
@@ -42,24 +43,15 @@ See the `src/test/resource/application.properties` to see the credentials used b
 
 ## Before run the application in dev mode
 
-### Keycloak and PostgreSQL  
+### Keycloak (20.0.1) and PostgreSQL (15.1)  
 
 This application depend on these services, so, you must have them running.
 
-For dev purpose, you can use the bash script `./start-keycloak.sh` to start Docker containers with these services. 
-The content of the script is below:
-```bash
-docker-compose -f src/main/docker/keycloak/docker-compose.yaml --env-file=src/main/docker/keycloak/.env build
-docker-compose -f src/main/docker/keycloak/docker-compose.yaml --env-file=src/main/docker/keycloak/.env up -d
-docker-compose -f src/main/docker/keycloak/docker-compose.yaml --env-file=src/main/docker/keycloak/.env logs -f
-```
+For dev purpose, you can use the bash script `./start-keycloak.sh` to start Docker containers with these services.
 
 The Keycloak will be available at `http://localhost:8080`.
 
-To stop the containers you can use the script `./stop-keycloak.sh`. The content of the script is show below:
-```bash
-docker-compose -f src/main/docker/keycloak/docker-compose.yaml --env-file=src/main/docker/keycloak/.env down
-```
+To stop the containers you can use the script `./stop-keycloak.sh`.
 
 ### Keycloak Realm
 
@@ -71,9 +63,9 @@ Double-check the following properties of the `src/main/resource/application.prop
 The values of them must match with the values used to create the test realm.
 
 ```
-auth-server-url=${QUARKUS_OIDC_AUTH-SERVER-URL:http://localhost:8080/auth/realms/app}
-quarkus.oidc.client-id=${QUARKUS_OIDC_CLIENT-ID:secure-api-quarkus}
-quarkus.oidc.credentials.secret=${QUARKUS_OIDC_CREDENTIALS_SECRET:rQGYWLVWPNd8brFCqUfodNHv1SUfZ6We}
+auth-server-url=${QUARKUS_OIDC_AUTH-SERVER-URL:http://localhost:8080/realms/app}
+quarkus.oidc.client-id=${QUARKUS_OIDC_CLIENT-ID:secure-api}
+quarkus.oidc.credentials.secret=${QUARKUS_OIDC_CREDENTIALS_SECRET:change_me}
 ```
 
 ### Run the application in dev mode
