--- src/main/docker/keycloak/Dockerfile-keycloak
@@ -1,2 +1,0 @@
-FROM quay.io/keycloak/keycloak:18.0.2-legacy
-COPY ./keycloak-add-user.json /opt/jboss/keycloak/standalone/configuration/keycloak-add-user.json
