+++ src/main/docker/keycloak/PostgreSQL.dockerfile
@@ -1,2 +1,2 @@
-FROM postgres:13.3
+FROM postgres:15.1
 COPY ./init-user-db.sh /docker-entrypoint-initdb.d/init-user-db.sh
