+++ src/main/docker/keycloak/docker-compose.env.example
@@ -1,0 +1,24 @@
+POSTGRES_DATA_DIR=/var/lib/postgresql-docker/data
+POSTGRES_USER=postgres
+POSTGRES_PASSWORD=postgres
+POSTGRES_HOST_AUTH_METHOD=password
+POSTGRES_INITDB_ARGS=--auth-host=password
+POSTGRES_PORT=127.0.0.1:5432:5432
+
+APP_DB_NAME=app_dev
+APP_DB_USERNAME=app_dev
+APP_DB_PASSWORD=change_me
+
+APP_DB_TEST_NAME=app_test
+APP_DB_TEST_USERNAME=app_test
+APP_DB_TEST_PASSWORD=change_me
+
+KC_PORT=127.0.0.1:8080:8080
+KC_DB_URL_HOST=postgresql
+KC_DB_URL_PORT=5432
+KC_DB_SCHEMA=public
+KC_DB_URL_DATABASE=keycloak
+KC_DB_USERNAME=keycloak
+KC_DB_PASSWORD=change_me
+KC_LOG_LEVEL=ERROR
+KEYCLOAK_ADMIN_PASSWORD=admin
