+++ src/main/docker/keycloak/docker-compose.yaml
@@ -1,11 +1,12 @@
 # Container only for development and testing! purpose
-version: "3.8"
+version: "3.7"
 services:
   postgresql:
-    container_name: postgresql
+    # https://hub.docker.com/_/postgres
     build:
       context: .
-      dockerfile: Dockerfile-postgresql
+      dockerfile: PostgreSQL.dockerfile
+    container_name: postgresql
     volumes:
       - ${POSTGRES_DATA_DIR}:/var/lib/postgresql/data/
     environment:
@@ -13,23 +14,35 @@ services:
       - POSTGRES_PASSWORD=${POSTGRES_PASSWORD}
       - POSTGRES_HOST_AUTH_METHOD=${POSTGRES_HOST_AUTH_METHOD}
       - POSTGRES_INITDB_ARGS=${POSTGRES_INITDB_ARGS}
-      - POSTGRES_DATABASES=${POSTGRES_DATABASES}
+      - APP_DB_NAME=${APP_DB_NAME}
+      - APP_DB_USERNAME=${APP_DB_USERNAME}
+      - APP_DB_PASSWORD=${APP_DB_PASSWORD}
+      - APP_DB_TEST_NAME=${APP_DB_TEST_NAME}
+      - APP_DB_TEST_USERNAME=${APP_DB_TEST_USERNAME}
+      - APP_DB_TEST_PASSWORD=${APP_DB_TEST_PASSWORD}
+      - KC_DB_URL_DATABASE=${KC_DB_URL_DATABASE}
+      - KC_DB_USERNAME=${KC_DB_USERNAME}
+      - KC_DB_PASSWORD=${KC_DB_PASSWORD}
     ports:
       - ${POSTGRES_PORT}
+  # https://quay.io/repository/keycloak/keycloak
   keycloak:
     container_name: keycloak
+    image: quay.io/keycloak/keycloak:20.0.1
+    command:
+      - start-dev
+    ports:
+      - ${KC_PORT}
+    environment:
+      - KC_DB=postgres
+      - KC_DB_URL_HOST=${KC_DB_URL_HOST}
+      - KC_DB_URL_PORT=${KC_DB_URL_PORT}
+      - KC_DB_SCHEMA=${KC_DB_SCHEMA}
+      - KC_DB_URL_DATABASE=${KC_DB_URL_DATABASE}
+      - KC_DB_USERNAME=${KC_DB_USERNAME}
+      - KC_DB_PASSWORD=${KC_DB_PASSWORD}
+      - KC_LOG_LEVEL=${KC_LOG_LEVEL}
+      - KEYCLOAK_ADMIN=admin
+      - KEYCLOAK_ADMIN_PASSWORD=${KEYCLOAK_ADMIN_PASSWORD}
     depends_on:
       - postgresql
-    build:
-      context: .
-      dockerfile: Dockerfile-keycloak
-    ports:
-      - "${KEYCLOAK_PORT}"
-    environment:
-      - DB_VENDOR=POSTGRES
-      - DB_ADDR=postgresql
-      - DB_SCHEMA=${KEYCLOAK_DB_SCHEMA}
-      - DB_DATABASE=${KEYCLOAK_DB_NAME}
-      - DB_USER=${KEYCLOAK_DB_USER}
-      - DB_PASSWORD=${KEYCLOAK_DB_PASSWORD}
-      - KEYCLOAK_LOGLEVEL=${KEYCLOAK_LOGLEVEL}
