+++ src/main/docker/keycloak/init-user-db.sh
@@ -1,20 +1,37 @@
 #!/bin/bash
 set -e
-IFS=';' databases=(${POSTGRES_DATABASES})
-for database in "${databases[@]}"
-do
-    IFS=',' credentials=(${database})
-    # create the user 
-    psql -v ON_ERROR_STOP=1 --username "$POSTGRES_USER" --dbname "$POSTGRES_DB" -c "
-        CREATE ROLE ${credentials[1]}
-        WITH LOGIN NOSUPERUSER NOINHERIT NOCREATEDB NOCREATEROLE NOREPLICATION
-        PASSWORD '${credentials[2]}'"
 
-    # create the database with the user as owner
-    psql -v ON_ERROR_STOP=1 --username "$POSTGRES_USER" --dbname "$POSTGRES_DB" -c "
-        CREATE DATABASE ${credentials[0]}
-        WITH
-        OWNER = ${credentials[1]}
-        ENCODING = 'UTF8'"
-done
+# app database for dev
+psql -v ON_ERROR_STOP=1 --username "$POSTGRES_USER" --dbname "$POSTGRES_DB" -c "
+    CREATE ROLE ${APP_DB_USERNAME}
+    WITH NOSUPERUSER NOCREATEDB NOCREATEROLE NOINHERIT LOGIN NOREPLICATION NOBYPASSRLS
+    PASSWORD '${APP_DB_PASSWORD}'"
+psql -v ON_ERROR_STOP=1 --username "$POSTGRES_USER" --dbname "$POSTGRES_DB" -c "
+    CREATE DATABASE ${APP_DB_NAME}
+    WITH OWNER = ${APP_DB_USERNAME} ENCODING = 'UTF8'"
+psql -v ON_ERROR_STOP=1 --username "$POSTGRES_USER" --dbname "$APP_DB_USERNAME" -c "
+    GRANT ALL ON SCHEMA public to ${APP_DB_NAME}"
+
+# app database for test (in case run the tests without Testcontainers)
+psql -v ON_ERROR_STOP=1 --username "$POSTGRES_USER" --dbname "$POSTGRES_DB" -c "
+    CREATE ROLE ${APP_DB_TEST_USERNAME}
+    WITH NOSUPERUSER NOCREATEDB NOCREATEROLE NOINHERIT LOGIN NOREPLICATION NOBYPASSRLS
+    PASSWORD '${APP_DB_TEST_PASSWORD}'"
+psql -v ON_ERROR_STOP=1 --username "$POSTGRES_USER" --dbname "$POSTGRES_DB" -c "
+    CREATE DATABASE ${APP_DB_TEST_NAME}
+    WITH OWNER = ${APP_DB_TEST_USERNAME} ENCODING = 'UTF8'"
+psql -v ON_ERROR_STOP=1 --username "$POSTGRES_USER" --dbname "$APP_DB_TEST_NAME" -c "
+    GRANT ALL ON SCHEMA public to ${APP_DB_TEST_USERNAME}"
+
+# Keycloak database
+psql -v ON_ERROR_STOP=1 --username "$POSTGRES_USER" --dbname "$POSTGRES_DB" -c "
+    CREATE ROLE ${KC_DB_USERNAME}
+    WITH NOSUPERUSER NOCREATEDB NOCREATEROLE NOINHERIT LOGIN NOREPLICATION NOBYPASSRLS
+    PASSWORD '${KC_DB_PASSWORD}'"
+psql -v ON_ERROR_STOP=1 --username "$POSTGRES_USER" --dbname "$POSTGRES_DB" -c "
+    CREATE DATABASE ${KC_DB_URL_DATABASE}
+    WITH OWNER = ${KC_DB_USERNAME} ENCODING = 'UTF8'"
+psql -v ON_ERROR_STOP=1 --username "$POSTGRES_USER" --dbname "$KC_DB_URL_DATABASE" -c "
+    GRANT ALL ON SCHEMA public to ${KC_DB_USERNAME}"
+
