--- src/main/docker/keycloak/keycloak-add-user.json
@@ -1,13 +1,0 @@
-[ {
-  "realm" : "master",
-  "users" : [ {
-    "username" : "admin",
-    "enabled" : true,
-    "credentials" : [ {
-      "type" : "password",
-      "secretData" : "{\"value\":\"34OeGqsDnieTbXEyWn4sHyatPWG00EyBA9cEv9H0AYVKdjyR8i86NrvPSsVwRbbBjyVlk36a7+BUltM5sWhWWw==\",\"salt\":\"dlgjpzd2I9IDCqd7aIAI6w==\",\"additionalParameters\":{}}",
-      "credentialData" : "{\"hashIterations\":100000,\"algorithm\":\"pbkdf2-sha256\",\"additionalParameters\":{}}"
-    } ],
-    "realmRoles" : [ "admin" ]
-  } ]
-} ]
