+++ src/main/java/com/mycompany/exceptions/WebApplicationExceptionMapper.java
@@ -1,0 +1,14 @@
+package com.mycompany.exceptions;
+
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.Response;
+import javax.ws.rs.ext.ExceptionMapper;
+import javax.ws.rs.ext.Provider;
+
+@Provider
+public class WebApplicationExceptionMapper implements ExceptionMapper<WebApplicationException> {
+    @Override
+    public Response toResponse(WebApplicationException e) {
+        return e.getResponse();
+    }
+}
