+++ src/main/postman/Local.postman_environment.json
@@ -1,0 +1,38 @@
+{
+	"id": "a57f57de-5d52-4bca-a41d-bdd1857437cc",
+	"name": "Local",
+	"values": [
+		{
+			"key": "keycloakUrl",
+			"value": "http://localhost:8080",
+			"enabled": true
+		},
+		{
+			"key": "apiUrl",
+			"value": "http://localhost:8081",
+			"type": "default",
+			"enabled": true
+		},
+		{
+			"key": "authToken",
+			"value": "",
+			"type": "default",
+			"enabled": true
+		},
+		{
+			"key": "keycloakClientSecret",
+			"value": "",
+			"type": "default",
+			"enabled": true
+		},
+		{
+			"key": "keycloakClientID",
+			"value": "secure-api",
+			"type": "default",
+			"enabled": true
+		}
+	],
+	"_postman_variable_scope": "environment",
+	"_postman_exported_at": "2022-10-19T06:44:50.702Z",
+	"_postman_exported_using": "Postman/9.31.21"
+}
