+++ src/main/resources/application.properties
@@ -11,16 +11,16 @@ quarkus.locales=en,pt-BR
 quarkus.oidc.tls.verification=none
 
 # Set the Keycloak Configuration (Copy the secret generated on Keycloak application)
-quarkus.oidc.auth-server-url=${QUARKUS_OIDC_AUTH-SERVER-URL:http://localhost:8080/auth/realms/app}
-quarkus.oidc.client-id=${QUARKUS_OIDC_CLIENT-ID:secure-api-quarkus}
-quarkus.oidc.credentials.secret=${QUARKUS_OIDC_CREDENTIALS_SECRET:rQGYWLVWPNd8brFCqUfodNHv1SUfZ6We}
+quarkus.oidc.auth-server-url=${QUARKUS_OIDC_AUTH-SERVER-URL:http://localhost:8080/realms/app}
+quarkus.oidc.client-id=${QUARKUS_OIDC_CLIENT-ID:secure-api}
+quarkus.oidc.credentials.secret=${QUARKUS_OIDC_CREDENTIALS_SECRET:7Is9GoZzWeIXkPC5Z67dAlPjeaHm9oMn}
 
 # Set the database configurations
 quarkus.datasource.db-kind=postgresql
 
-quarkus.datasource.jdbc.url=${QUARKUS_DATASOURCE_JDBC_URL:jdbc:postgresql://localhost:5432/app}
-quarkus.datasource.username=${QUARKUS_DATASOURCE_USERNAME:app_user}
-quarkus.datasource.password=${QUARKUS_DATASOURCE_PASSWORD:password}
+quarkus.datasource.jdbc.url=${QUARKUS_DATASOURCE_JDBC_URL:jdbc:postgresql://localhost:5432/app_dev}
+quarkus.datasource.username=${QUARKUS_DATASOURCE_USERNAME:app_dev}
+quarkus.datasource.password=${QUARKUS_DATASOURCE_PASSWORD:change_me}
 
 # Disable the default Quarkus dev services as we created the services on our own
 quarkus.datasource.devservices.enabled=false
