+++ src/test/resources/application.properties
@@ -1,8 +1,10 @@
 quarkus.http.test-port=8082
 
 quarkus.datasource.jdbc.url=jdbc:postgresql://localhost:5432/app_test
-quarkus.datasource.username=app_test_user
-quarkus.datasource.password=password
+quarkus.datasource.username=app_test
+quarkus.datasource.password=change_me
+
+quarkus.oidc.enabled=false
 
 # Drop and recreate the database every time the tests run
 quarkus.hibernate-orm.database.generation=drop-and-create
