+++ start-keycloak.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
-docker-compose -f src/main/docker/keycloak/docker-compose.yaml --env-file=src/main/docker/keycloak/.env build
-docker-compose -f src/main/docker/keycloak/docker-compose.yaml --env-file=src/main/docker/keycloak/.env up -d
-docker-compose -f src/main/docker/keycloak/docker-compose.yaml --env-file=src/main/docker/keycloak/.env logs -f
+if [ ! -f "./src/main/docker/keycloak/docker-compose.env" ]
+then
+  cp ./src/main/docker/keycloak/docker-compose.env.example ./src/main/docker/keycloak/docker-compose.env
+fi
+docker-compose -f ./src/main/docker/keycloak/docker-compose.yaml --env-file=./src/main/docker/keycloak/docker-compose.env build
+docker-compose -f ./src/main/docker/keycloak/docker-compose.yaml --env-file=./src/main/docker/keycloak/docker-compose.env up -d
+docker-compose -f ./src/main/docker/keycloak/docker-compose.yaml --env-file=./src/main/docker/keycloak/docker-compose.env logs -f
